+++ supreme/verifier/src/jvmTest/kotlin/examples/Readme-Verifier-min.kt
@@ -1,4 +1,6 @@
 package examples.docs.config.minimal
 import at.asitplus.attestation.supreme.AttestationVerifier
 
+// --8<-- [start:minimal-verifier]
 val verifier = AttestationVerifier(makoto)/*(1)!*/
+// --8<-- [end:minimal-verifier]